adding image trait

+++ app/Helpers/Helpers.php
@@ -4,26 +4,32 @@ namespace App\Helpers;
 
 use Illuminate\Support\Facades\Exceptions;
 use Illuminate\Support\Facades\Storage;
+use Intervention\Image\Drivers\Imagick\Driver;
+use Intervention\Image\ImageManager;
 
 class Helpers
 {
-    public function makeDirectory(): string
+    public function saveImage($image): string
     {
-        $subDir = 'images/' . date('Y/m/d');
-        Storage::makeDirectory($subDir);
-        return $subDir;
-    }
-
-    public function uploadImage($image): string
-    {
-        $path = $this->makeDirectory();
-        $image->store($path);
-        return $path;
+        $directory = $this->makeDirectory();
+        $manager = new ImageManager(new Driver());
+        $manager->read($image)->toWebp(90)->save($directory . '/' . time() . '_' . rand(10, 1000) . '_image.webp');
+        return $directory . '/' . time() . '_' . rand(10, 1000) . '_image.webp';
     }
 
     public function removeImage($image): void
     {
-        Storage::delete($image);
+        if ($image !== null) {
+            Storage::delete($image);
+        }
+
+    }
+
+    private static function makeDirectory(): string
+    {
+        $subFolder = 'images/' . date('Y/m/d');
+        Storage::makeDirectory($subFolder);
+        return $subFolder;
     }
 
 }
